Bump Python dependencies and adapt to pandas/SQLAlchemy 2.x.
Pin current package releases, move to sqlalchemy.orm.declarative_base,
and fix pivot-table dtype and seaborn/pandas API nits exposed by the bump.

+++ analysis/test_data_utils.py
@@ -354,7 +354,7 @@ def test_experiment_pivot_table():
     expected_pivot_table = pd.pivot_table(expected_data,
                                           index=['benchmark'],
                                           columns=['fuzzer'],
-                                          values='median')
+                                          values='median').astype(int)
     assert pivot_table.equals(expected_pivot_table)
